Fix scrollbar behavior on history load
If the bar is at the bottom and a new message comes,
scroll down to see it.

If it's not at the bottom (i.e. user is looking at previous messages)
but not a the top either and a new message comes, don't scroll.

If it hits the top, reposition it for every message prepended,
fast enough so that it appears to not move.