Use pure signals instead of queues for messages
Chats will be intialized under the hood when messages arrive on start.
When user clicks a chat, the same initialized chats are displayed
without lag thanks to the cache.

Inserted message dates are kept track of with a deque, so past messages
will always be inserted at the right position.